mdl_order has been changed from int to size_t
git-svn-id: file:///home/roberto/Git/svn@8205 1a721e8d-d38c-4e91-ab98-6a0be4d754af

Former-commit-id: 689afe15a2cdc064c4a95d555e3d057ac3a25c3f

+++ trunk/xmipp/libraries/reconstruction/ctf_group.cpp
@@ -256,7 +256,7 @@ void ProgCtfGroup::produceSideInfo()
 
     //fill multiarray with ctfs
     int count;
-    int counter=0;
+    size_t counter=0;
     if (verbose!=0)
         std::cout << "\nFill multiarray with ctfs" <<std::endl;
     FOR_ALL_OBJECTS_IN_METADATA(ctfMD)
@@ -293,7 +293,7 @@ void ProgCtfGroup::produceSideInfo()
                 FOR_ALL_DIRECT_ELEMENTS_IN_ARRAY2D(Mctf)
                 {
                     id=md1.addObject();
-                    md1.setValue(MDL_ORDER,(int)(id-1),id);
+                    md1.setValue(MDL_ORDER,(size_t)(id-1),id);
                     md1.setValue(MDL_RESOLUTION_FRC,dAij(Mctf, 0, j),id );
                 }
                 std::stringstream ss;
